Moved Swagger config to his own config file
